# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/jobs/offers": {
      "post": {
        "operationId": "JobsController_createOffer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobOfferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      },
      "get": {
        "operationId": "JobsController_findAllOffers",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sector",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
...
    "/api/jobs/offers/{id}": {
      "get": {
        "operationId": "JobsController_findOneOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      },
      "patch": {
        "operationId": "JobsController_updateOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobOfferDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      },
      "delete": {
        "operationId": "JobsController_deleteOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/offers/{id}/status": {
      "patch": {
        "operationId": "JobsController_updateOfferStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/apply": {
      "post": {
        "operationId": "JobsController_apply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/apply/auth": {
      "post": {
        "operationId": "JobsController_applyAuth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/offers/{id}/applications": {
      "get": {
        "operationId": "JobsController_getApplicationsByOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/applications/{id}": {
      "get": {
        "operationId": "JobsController_getApplicationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/applications/{id}/status": {
      "patch": {
        "operationId": "JobsController_updateApplicationStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Jobs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Y-Mad API",
    "description": "\n      API complete du site Y-Mad (Youthful Madagascar)\n\n      INFORMATIONS DE CONTACT:\n      - Adresse: Carion, Antananarivo, Madagascar\n      - Telephone: +261 32 04 856 97\n      - Email: ymad.mg@gmail.com\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Y-Mad Association",
      "url": "https://y-mad.mg",
      "email": "ymad.mg@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Entrez votre token JWT"
      }
    },
    "schemas": {
...
      "CreateJobOfferDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateJobOfferDto": {
        "type": "object",
        "properties": {}
      },
      "CreateJobApplicationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}